third_party/blink/web_tests/external/wpt/fetch/data-urls/processing.any.sharedworker-expected.txt

This is a testharness.js-based test.
Found 18 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] "data:;x=x;charset=x,X"
  assert_equals: expected "text/plain;x=x;charset=x" but got "text/plain;charset=x"
[FAIL] "data:;x=x,X"
  assert_equals: expected "text/plain;x=x" but got "text/plain;charset=US-ASCII"
[FAIL] "data:IMAGE/gif;hi=x,%C2%B1"
  assert_equals: expected "image/gif;hi=x" but got "image/gif"
[FAIL] "data:text/plain;a=\\",\\",X"
  assert_equals: expected "text/plain;a=\\"\\"" but got "text/plain"
[FAIL] "data:text/plain;a=%2C,X"
  assert_equals: expected "text/plain;a=%2C" but got "text/plain"
[FAIL] "data:;base64;base64,WA"
  assert_equals: expected "text/plain" but got "text/plain;charset=US-ASCII"
[FAIL] "data:x/x;base64;charset=x,WA"
  assert_array_equals: lengths differ, expected array [87, 65] length 2, got object "88" length 1
[FAIL] "data:x/x;base64;base64x,WA"
  assert_array_equals: lengths differ, expected array [87, 65] length 2, got object "88" length 1
[FAIL] "data:x;base64;x,WA"
  assert_array_equals: lengths differ, expected array [87, 65] length 2, got object "88" length 1
[FAIL] "data:;base64;,WA"
  assert_array_equals: lengths differ, expected array [87, 65] length 2, got object "88" length 1
[FAIL] "data:;base 64,WA"
  assert_equals: expected "text/plain" but got "text/plain;charset=US-ASCII"
[FAIL] "data:;%62ase64,WA"
  assert_equals: expected "text/plain" but got "text/plain;charset=US-ASCII"
[FAIL] "data:;charset =x,X"
  assert_equals: expected "text/plain" but got "text/plain;charset=US-ASCII"
[FAIL] "data:;charset= x,X"
  promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
[FAIL] "data:;charset=,X"
  promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
[FAIL] "data:;charset,X"
  assert_equals: expected "text/plain" but got "text/plain;charset=US-ASCII"
[FAIL] "data:;charset=\\"x\\",X"
  promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
[FAIL] "data:;CHARSET=\\"X\\",X"
  promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
Harness: the test ran to completion.
